+++ tests/test_public_api.py
@@ -33,6 +33,7 @@ def test_parse_path_simple():
     assert p == "proj"
     assert r == "run"
 
+
 def test_parse_path_leading():
     u, p, r = api._parse_path("/user/proj/run")
     assert u == "user"
@@ -189,6 +190,17 @@ def test_runs_from_path_index(mocker, request_mocker, query_runs_v2, query_downl
     run_mock.side_effect = [True, False]
     assert runs[3]
     assert len(runs.objects) == 4
+
+
+def test_projects(mocker, request_mocker, query_projects_v2):
+    runs_mock = query_projects_v2(request_mocker)
+    projects = api.projects("test")
+    # projects doesn't provide a length for now, so we iterate
+    # them all to count
+    count = 0
+    for proj in projects:
+        count += 1
+    assert count == 2
 
 
 # @pytest.mark.skip(readon='fails when I run the whole suite, but not when I run just this file')